+++ RELEASE_NOTES.md
@@ -1,10 +1,10 @@
 ## VPinFE Release Notes
 
 ### Summary
-This is the offical release of vpinfe v1.1.1.  TESTING
+This is the offical release of vpinfe v1.1.1  TESTING ONLY
 
 ### What's New
-None
+- Clean up and revamp ManagerUI Configuration page. 
 
 ### Fixes
 None